Rest

+++ src/main/java/tn/esprit/pibakcend/repository/CommentaireRepository.java
@@ -3,5 +3,5 @@ package tn.esprit.pibakcend.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import tn.esprit.pibakcend.entities.Commentaire;
 
-public interface CommentaireRepository extends JpaRepository<Commentaire,Integer> {
+public interface CommentaireRepository extends JpaRepository<Commentaire,Long> {
 }
